fix(#273): unify AgentId as u64 across QuarantineRegistry, reconciler, and all boundaries (#297)
Promotes `phantom-agents::agent::AgentId` from `u32` to `u64`, matching
`phantom-agents::role::AgentId` and `QuarantineRegistry`. Removes the
`u32::try_from(agent_id).unwrap_or(u32::MAX)` saturating cast in the
reconciler (and the accompanying `agent_id as u32` cast in check_stalled).

Changes across the stack:
- phantom-agents/src/agent.rs: AgentId = u64 (was u32)
- phantom-brain/src/orchestrator.rs: PlanStep::agent_id Option<u64>
- phantom-brain/src/reconciler.rs: remove both narrowing casts; rewrite
  the #272 saturation test as a #273 round-trip correctness test
- phantom-protocol/src/events.rs: AgentId = u64 (IPC wire type)
- phantom-app/src/adapters/agent.rs: widen app_id to u64 at wire boundary
- phantom-app/src/agent_pane.rs: remove stale `as u64` casts, update comment
- phantom-session/src/{agent_state,goal_state}.rs: widen persistence types
- phantom-agents/src/failure_store.rs: fix test range literals (u32→u64)
- phantom-brain/src/brain.rs: fix test literal types (u32→u64)

cargo check --workspace passes; phantom-agents and phantom-brain test
suites pass (605 + 343 tests, 0 failures).

+++ crates/phantom-brain/src/orchestrator.rs
@@ -102,7 +102,10 @@ pub struct PlanStep {
     /// Current execution status.
     pub status: StepStatus,
     /// The agent ID if one has been spawned for this step.
-    pub agent_id: Option<u32>,
+    ///
+    /// Stored as `u64` to match [`phantom_agents::AgentId`] (fixes #273 — no
+    /// narrowing cast required at the reconciler / quarantine boundary).
+    pub agent_id: Option<u64>,
     /// Number of times this step has been attempted (for retry tracking).
     pub attempts: u32,
     /// Maximum attempts before marking as Failed.
